+++ src/Config.h
@@ -7,14 +7,13 @@ namespace lo {
 using ll::reflection::Dispatcher;
 struct Config {
 
-    int version = 17;
+    int version = 18;
 
     struct {
-        Dispatcher<bool, moving_block_opt::MovingBlockOpt>                        optMovingBlock  = true;
-        Dispatcher<bool, hopper_item_opt::HopperItemOpt>                          optHopperItem   = true;
-        Dispatcher<bool, chunk_leak_fix::ChunkLeakFix>                            fixChunkLeak    = true;
-        Dispatcher<bool, seen_percent_opt::SeenPercentOpt>                        optSeenPercent  = false;
-        Dispatcher<push_entity_opt::Config, push_entity_opt::PushEntityOpt>       optPushEntity{};
+        Dispatcher<bool, moving_block_opt::MovingBlockOpt>                  optMovingBlock = false;
+        Dispatcher<bool, chunk_leak_fix::ChunkLeakFix>                      fixChunkLeak   = true;
+        Dispatcher<bool, seen_percent_opt::SeenPercentOpt>                  optSeenPercent = false;
+        Dispatcher<push_entity_opt::Config, push_entity_opt::PushEntityOpt> optPushEntity{};
     } features{};
     struct {
         bool timingCommand = true;