+++ xmake.lua
@@ -1,18 +1,6 @@
-add_rules("mode.release", "mode.debug")
+add_rules("mode.debug", "mode.release")
 
-add_repositories("liteldev-repo https://github.com/LiteLDev/xmake-repo.git")
-
-if is_config("target_type", "server") then
-    add_requires("levilamina 26.10.*", {configs = {target_type = "server"}})
-else
-    add_requires("levilamina 26.10.*", {configs = {target_type = "client"}})
-end
-
-add_requires("levibuildscript")
-
-if not has_config("vs_runtime") then
-    set_runtimes("MD")
-end
+add_repositories("levimc-repo https://github.com/LiteLDev/xmake-repo.git")
 
 option("target_type")
     set_default("server")
@@ -20,43 +8,47 @@ option("target_type")
     set_values("server", "client")
 option_end()
 
+add_requires("levilamina 26.20.*", {configs = {target_type = get_config("target_type")}})
+
+add_requires("levibuildscript")
+
+if not has_config("vs_runtime") then
+    set_runtimes("MD")
+end
 
 target("LeviOptimize")
     add_rules("@levibuildscript/linkrule")
     add_rules("@levibuildscript/modpacker")
-    add_cxflags(
-        "/EHa",
-        "/utf-8",
-        "/W4",
-        "/w44265",
-        "/w44289",
-        "/w44296",
-        "/w45263",
-        "/w44738",
-        "/w45204"
-    )
-    add_defines(
-        "_HAS_CXX23=1", -- To enable C++23 features
-        "NOMINMAX", -- To avoid conflicts with std::min and std::max.
-        "UNICODE" -- To enable Unicode support in Windows API.
-    )
-    add_files(
-        "src/**.cpp"
-    )
-    add_includedirs(
-        "src"
-    )
+    if is_plat("windows") then
+        add_defines("NOMINMAX", "UNICODE")
+        set_exceptions("none") -- To avoid conflicts with /EHa.
+        add_cxflags( "/EHa", "/utf-8", "/W4", "/w44265", "/w44289", "/w44296", "/w45263", "/w44738", "/w45204")
+        add_cxflags(
+            "/EHs",
+            "-Wno-microsoft-cast",
+            "-Wno-invalid-offsetof",
+            "-Wno-c++2b-extensions",
+            "-Wno-microsoft-include",
+            "-Wno-overloaded-virtual",
+            "-Wno-ignored-qualifiers",
+            "-Wno-missing-field-initializers",
+            "-Wno-potentially-evaluated-expression",
+            "-Wno-pragma-system-header-outside-header",
+            {tools = {"clang_cl"}}
+        )
+        set_toolchains("clang-cl")
+    end
     add_packages("levilamina")
-    set_exceptions("none") -- To avoid conflicts with /EHa.
     set_kind("shared")
     set_languages("c++20")
     set_symbols("debug")
+    add_headerfiles("src/**.h")
+    add_files("src/**.cpp")
+    add_includedirs("src")
     if is_config("target_type", "server") then
-        add_defines("LL_PLAT_S")
     --  add_includedirs("src-server")
     --  add_files("src-server/**.cpp")
     else
-        add_defines("LL_PLAT_C")
     --  add_includedirs("src-client")
     --  add_files("src-client/**.cpp")
     end