+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.14.0] - 2026-08-27
+
+### Changed
+
+- Adapted to LeviLamina 26.20.*
+- Removed hopper item optimization
+
 ## [0.13.2] - 2026-05-19
 
 ### Fixed
@@ -209,7 +216,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 [#26]: https://github.com/LiteLDev/LeviOptimize/issues/26
 
-[Unreleased]: https://github.com/LiteLDev/LeviOptimize/compare/v0.13.2...HEAD
+[Unreleased]: https://github.com/LiteLDev/LeviOptimize/compare/v0.14.0...HEAD
+[0.14.0]: https://github.com/LiteLDev/LeviOptimize/compare/v0.13.2...v0.14.0
 [0.13.2]: https://github.com/LiteLDev/LeviOptimize/compare/v0.13.1...v0.13.2
 [0.13.1]: https://github.com/LiteLDev/LeviOptimize/compare/v0.13.0...v0.13.1
 [0.13.0]: https://github.com/LiteLDev/LeviOptimize/compare/v0.12.1...v0.13.0